+++ src/pages/board/index.tsx
@@ -1,11 +1,13 @@
 import React, { useMemo, useState, MouseEvent } from 'react'
 import { PaintBoard } from '@/utils/paintBoard'
 import { CANVAS_ELE_TYPE } from '@/utils/constants'
-import ToolPanel from '../../components/toolPanel'
 import { useBackspace, useResizeEvent, useSpaceEvent } from '@/hooks/event'
-import Info from '../../components/info'
 import { CURSOR_TYPE } from '@/utils/cursor'
 import { TextEdit } from '@/utils/element/text'
+
+import ToolPanel from '@/components/toolPanel'
+import Info from '@/components/info'
+import CleanModal from '@/components/cleanModal'
 
 const textEdit = new TextEdit()
 
@@ -154,6 +156,7 @@ const Board: React.FC = () => {
         onDoubleClick={dbClick}
       ></canvas>
       <Info />
+      <CleanModal board={board} />
     </div>
   )
 }